RbConfig Initialization
While building the jekyll static site locally, advisory notice was given in the command window stating I should add the above to avoid always having to poll for changes. Figure it has something to do with me running this on windows. Possibly make it a bit tricky to collaboratively develop blog as each fork of this repo will be hosted on various OSes. Will refer once tested for efficacy.